NEXUS-46321: add policy evaluation step
In NEXUS-46138 we needed to significantly refactor the publishing job. That job previously carried an IQ policy evaluation that no longer fits in how that job needs to work to achieve its goals. Restoring the policy evaluation step in the prepare phase will ensure its present for the later SBOM collection.